+++ test/units/map-early-response.test.ts
@@ -54,7 +54,7 @@ describe('Map Early Response', () => {
 		const response = mapEarlyResponse(body, defaultContext)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response?.json<any>()).toEqual(body)
+		expect(await response?.json()).toEqual(body)
 		expect(response?.status).toBe(200)
 	})
 
@@ -87,7 +87,7 @@ describe('Map Early Response', () => {
 		)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response?.json<any>()).toEqual(body)
+		expect(await response?.json()).toEqual(body)
 		expect(response?.status).toBe(200)
 	})
 
@@ -106,7 +106,7 @@ describe('Map Early Response', () => {
 		const response = mapEarlyResponse(new Student('Himari'), defaultContext)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response?.json<any>()).toEqual({
+		expect(await response?.json()).toEqual({
 			name: 'Himari'
 		})
 		expect(response?.status).toBe(200)
@@ -143,7 +143,7 @@ describe('Map Early Response', () => {
 		)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response?.json<any>()).toEqual(body)
+		expect(await response?.json()).toEqual(body)
 		expect(response?.headers.toJSON()).toEqual({
 			...context.headers,
 			'content-type': 'application/json;charset=utf-8'
@@ -155,7 +155,7 @@ describe('Map Early Response', () => {
 		const response = mapEarlyResponse(new Error('Hello'), context)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response?.json<any>()).toEqual({
+		expect(await response?.json()).toEqual({
 			name: 'Error',
 			message: 'Hello'
 		})