+++ test/units/map-response.test.ts
@@ -56,7 +56,7 @@ describe('Map Response', () => {
 		const response = mapResponse(body, defaultContext)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response.json<any>()).toEqual(body)
+		expect(await response.json()).toEqual(body)
 		expect(response.status).toBe(200)
 	})
 
@@ -105,7 +105,7 @@ describe('Map Response', () => {
 		)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response.json<any>()).toEqual(body)
+		expect(await response.json()).toEqual(body)
 		expect(response.status).toBe(200)
 	})
 
@@ -113,7 +113,7 @@ describe('Map Response', () => {
 		const response = mapResponse(new Error('Hello'), defaultContext)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response.json<any>()).toEqual({
+		expect(await response.json()).toEqual({
 			name: 'Error',
 			message: 'Hello'
 		})
@@ -140,7 +140,7 @@ describe('Map Response', () => {
 		const response = mapResponse(new Student('Himari'), defaultContext)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response.json<any>()).toEqual({
+		expect(await response.json()).toEqual({
 			name: 'Himari'
 		})
 		expect(response.status).toBe(200)
@@ -195,7 +195,7 @@ describe('Map Response', () => {
 		)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response.json<any>()).toEqual(body)
+		expect(await response.json()).toEqual(body)
 		expect(response.headers.toJSON()).toEqual({
 			...context.headers,
 			'content-type': 'application/json;charset=utf-8'
@@ -207,7 +207,7 @@ describe('Map Response', () => {
 		const response = mapResponse(new Error('Hello'), context)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response.json<any>()).toEqual({
+		expect(await response.json()).toEqual({
 			name: 'Error',
 			message: 'Hello'
 		})